split message and control messages.
